requirements.txt
azure-functions>=1.11.0
azure-storage-blob>=12.14.0
azure-core>=1.24.0